requirements.txt
Flask==2.3.3
Flask-RESTX==1.3.0
google-generativeai==0.5.4
python-dotenv==1.0.0
flask-cors==4.0.0 